maint: changed int to Py_ssize_t as suggested in mailing lists

+++ sisl/_indices.pxd
@@ -1,3 +1,3 @@
 # Define the interfaces for the functions exposed through cimport
 cdef int in_1d(const int[::1] array, const int v) nogil
-cdef int _index_sorted(const int[::1] array, const int v) nogil
+cdef Py_ssize_t _index_sorted(const int[::1] array, const int v) nogil
